# openapi.yaml
openapi: 3.0.3
info:
  title: Ordryn / GoTodo API
  description: |
    JSON API for Ordryn (GoTodo). Self-host with `GOTODO_MODE=full` (UI + API) or
    `GOTODO_MODE=api` (API only). Redis is required for `/api/v2` authentication,
    rate limiting, and live SSE fan-out.

    Auth: session cookie (SPA login/register) or `Authorization: Bearer <api_key>`.

    Rate limits (per user, Redis token bucket). Reads and writes are independent.
    Bearer API keys: 120 GET/min (~2/sec) and 60 writes/min (~1/sec).
    SPA session cookies use a higher budget (600 GET/min, 240 writes/min) so the
    web UI is not gated by the external API ceiling.

    Canonical paths are `/api/v2`. `/api/v1` remains as a compatibility alias
    for existing webhooks, OAuth callbacks, and clients.
  version: 2.0.0
  license:
    name: See LICENSE in repository
servers:
  - url: /
    description: Same-origin (BASE_PATH may prefix these paths)
...
paths:
...
  /api/v2/join-requests:
    post:
      tags: [Auth]
      summary: Submit a request to join the site
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
                message:
                  type: string
                  maxLength: 500
      responses:
        "200":
          description: Acknowledgement (same message even if already pending)
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  message:
                    type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/APIError"
        "429":
          $ref: "#/components/responses/RateLimited"
...
  /api/v2/me/extensions:
    get:
      tags: [Extensions]
      summary: List personal-inbox hook destinations
      description: Only site-enabled extensions that declare member-scoped settings are returned.
      security:
        - SessionCookie: []
        - BearerAuth: []
      responses:
        "200":
          description: Personal destinations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectExtensionsList"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v2/me/extensions/{extensionId}:
    parameters:
      - name: extensionId
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [Extensions]
      summary: Get a personal-inbox destination
      security:
        - SessionCookie: []
        - BearerAuth: []
      responses:
        "200":
          description: Personal destination
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectExtension"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Extensions]
      summary: Update a personal-inbox destination
      security:
        - SessionCookie: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectExtensionPatch"
      responses:
        "200":
          description: Updated destination
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectExtension"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v2/webhooks/inbound:
    post:
      tags: [Webhooks]
      summary: Generic inbound webhook (create, comment, complete, or set_field)
      description: Public HMAC or shared-secret receiver. Returns 403 until an admin enables inbound webhooks. complete and set_field require an enabled extension on the project that declares those actions. project_id is required. HMAC must include X-Ordryn-Timestamp (unix seconds, ±5 minutes); the signature is HMAC-SHA256 of timestamp + "." + raw body.
      security: []
      parameters:
        - name: X-Ordryn-Webhook-Secret
          in: header
          schema:
            type: string
        - name: X-Ordryn-Signature
          in: header
          description: HMAC SHA-256 of timestamp + "." + raw body, prefixed with sha256=
          schema:
            type: string
        - name: X-Ordryn-Timestamp
          in: header
          description: Unix seconds. Required for HMAC. Rejected if skewed more than 5 minutes.
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action, project_id]
              properties:
                action:
                  type: string
                  enum: [create, comment, complete, set_field]
                title:
                  type: string
                description:
                  type: string
                task_id:
                  type: integer
                comment:
                  type: string
                field:
                  type: string
                  description: Custom field key for set_field (extension_id.local_key).
                value:
                  type: string
                project_id:
                  type: integer
                  description: Required. Identifies the inbound project.
      responses:
        "200":
          description: Accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OK"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/APIError"
...
  /api/v2/projects/{id}/extensions:
    parameters:
      - $ref: "#/components/parameters/ID"
    get:
      tags: [Projects]
      summary: List site-enabled project-scoped extensions for members
      description: Only extensions enabled in Admin → Extensions and that declare a project surface are returned.
      security:
        - SessionCookie: []
        - BearerAuth: []
      responses:
        "200":
          description: Site-enabled extensions with this project's settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectExtensionsList"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v2/projects/{id}/extensions/{extensionId}:
    parameters:
      - $ref: "#/components/parameters/ID"
      - name: extensionId
        in: path
        required: true
        schema:
          type: string
    patch:
      tags: [Projects]
      summary: Update this project's extension settings (owner)
      security:
        - SessionCookie: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectExtensionPatch"
      responses:
        "200":
          description: Updated project extension
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectExtension"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v2/admin/extensions:
    get:
      tags: [Admin]
      summary: List loaded extensions
      security:
        - SessionCookie: []
        - BearerAuth: []
      responses:
        "200":
          description: Extensions with site-level enable state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminExtensionsList"
  /api/v2/admin/extensions/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [Admin]
      summary: Get one extension and its settings
      security:
        - SessionCookie: []
        - BearerAuth: []
      responses:
        "200":
          description: Extension
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminExtension"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Admin]
      summary: Enable or disable an extension for the site
      security:
        - SessionCookie: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminExtensionPatch"
      responses:
        "200":
          description: Updated extension
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminExtension"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v2/admin/join-requests:
    get:
      tags: [Admin]
      summary: List join requests
      security:
        - SessionCookie: []
        - BearerAuth: []
      responses:
        "200":
          description: Join requests (pending first)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JoinRequest"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API key
      description: API key from profile, bootstrap, or device auth
    SessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: httpOnly session cookie from /api/v2/auth/login or register
  parameters:
    ID:
      name: id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    RateLimited:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    APIError:
      description: API error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Machine-readable code (e.g. invalid_request)
        message:
          type: string
    OK:
      type: object
      properties:
        ok:
          type: boolean
...
    JoinRequest:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        message:
          type: string
        status:
          type: string
          enum: [pending, approved, denied]
        created_at:
          type: string
          format: date-time
        invite_id:
          type: integer
        reviewed_at:
          type: string
          format: date-time
        reviewed_by:
          type: integer
        invite_token:
          type: string
...
    AdminExtensionSettings:
      type: object
      properties:
        enabled:
          type: boolean
        triggers:
          type: array
          items:
            type: string
        templates:
          type: object
          additionalProperties:
            type: string
        last_error:
          type: string
        last_delivery_at:
          type: string
    AdminExtension:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        version:
          type: string
        host_api:
          type: integer
        status:
          type: string
          enum: [loaded, failed]
        error:
          type: string
        manifest:
          type: object
          additionalProperties: true
        settings:
          $ref: "#/components/schemas/AdminExtensionSettings"
        secrets:
          type: object
          additionalProperties:
            type: boolean
        deliveries:
          type: array
          items:
            $ref: "#/components/schemas/ExtensionDelivery"
    AdminExtensionsList:
      type: object
      properties:
        extensions:
          type: array
          items:
            $ref: "#/components/schemas/AdminExtension"
    AdminExtensionPatch:
      type: object
      properties:
        enabled:
          type: boolean
        webhook_url:
          type: string
        triggers:
          type: array
          items:
            type: string
        templates:
          type: object
          additionalProperties:
            type: string
    ProjectExtensionSettings:
      type: object
      properties:
        enabled:
          type: boolean
        triggers:
          type: array
          items:
            type: string
        templates:
          type: object
          additionalProperties:
            type: string
        status_only:
          type: boolean
        skip_self:
          type: boolean
        min_priority:
          type: integer
        tag_ids:
          type: array
          items:
            type: integer
        status_ids:
          type: array
          items:
            type: integer
        status_exclude_ids:
          type: array
          items:
            type: integer
        claimed_only:
          type: boolean
        claimed_is_me:
          type: boolean
        field_key:
          type: string
        field_value:
          type: string
        quiet_hours_start:
          type: string
        quiet_hours_end:
          type: string
        digest:
          type: string
        mention_map:
          type: object
          additionalProperties:
            type: string
        values:
          type: object
          additionalProperties:
            type: string
        last_error:
          type: string
        last_delivery_at:
          type: string
    ProjectExtension:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        version:
          type: string
        host_api:
          type: integer
        site_enabled:
          type: boolean
        manifest:
          type: object
          additionalProperties: true
        settings:
          $ref: "#/components/schemas/ProjectExtensionSettings"
        secrets:
          type: object
          additionalProperties:
            type: boolean
        member:
          $ref: "#/components/schemas/ProjectExtensionSettings"
        member_secrets:
          type: object
          additionalProperties:
            type: boolean
        signing_set:
          type: boolean
        member_signing_set:
          type: boolean
        signing_secret:
          type: string
          description: Plaintext HMAC secret, shown only immediately after rotate_signing.
        sample_json:
          type: string
        deliveries:
          type: array
          items:
            $ref: "#/components/schemas/ExtensionDelivery"
        member_deliveries:
          type: array
          items:
            $ref: "#/components/schemas/ExtensionDelivery"
    ProjectExtensionsList:
      type: object
      properties:
        extensions:
          type: array
          items:
            $ref: "#/components/schemas/ProjectExtension"
        is_owner:
          type: boolean
    ProjectExtensionPatch:
      type: object
      properties:
        enabled:
          type: boolean
        triggers:
          type: array
          items:
            type: string
        templates:
          type: object
          additionalProperties:
            type: string
        status_only:
          type: boolean
        skip_self:
          type: boolean
        min_priority:
          type: integer
        tag_ids:
          type: array
          items:
            type: integer
        status_ids:
          type: array
          items:
            type: integer
        status_exclude_ids:
          type: array
          items:
            type: integer
        claimed_only:
          type: boolean
        claimed_is_me:
          type: boolean
        field_key:
          type: string
        field_value:
          type: string
        quiet_hours_start:
          type: string
        quiet_hours_end:
          type: string
        digest:
          type: string
          enum: ["", hourly, daily]
        mention_map:
          type: object
          additionalProperties:
            type: string
        webhook_url:
          type: string
          description: Write-only incoming webhook URL. Omit or blank to keep the stored secret. Allowed hosts depend on the extension delivery type (Discord, Slack, Teams, ntfy, or generic HTTPS).
        ntfy_auth:
          type: string
        rotate_signing:
          type: boolean
        rotate_callback:
          type: boolean
        values:
          type: object
          additionalProperties:
            type: string
    ExtensionDelivery:
      type: object
      properties:
        id:
          type: integer
          format: int64
        event:
          type: string
        event_id:
          type: string
        host:
          type: string
        status:
          type: string
        http_code:
          type: integer
        error:
          type: string
        attempts:
          type: integer
        created_at:
          type: string
          format: date-time
...